requirements.txt
fastapi
uvicorn
pymongo
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
python-dotenv
